main: Move kim-specific code to new file
This should make it easier to reason about.
