+++ background.ts
@@ -3,18 +3,18 @@ import fs from 'fs';
 import path from 'path';
 import os from 'os';
 import { URL } from 'url';
-import util from 'util';
 
 import Electron from 'electron';
 import _ from 'lodash';
 import MacCA from 'mac-ca';
 import WinCA from 'win-ca';
 
+import mainEvents from '@/main/mainEvents';
+import { setupKim } from '@/main/kim';
 import * as settings from './src/config/settings';
 import { Tray } from './src/menu/tray.js';
 import window from './src/window/window.js';
 import * as K8s from './src/k8s-engine/k8s';
-import Kim from './src/k8s-engine/kim';
 import resources from './src/resources';
 import Logging from './src/utils/logging';
 import * as childProcess from './src/utils/childProcess';
@@ -24,11 +24,9 @@ Electron.app.setName('Rancher Desktop');
 const console = new Console(Logging.background.stream);
 
 let k8smanager: K8s.KubernetesBackend;
-let imageManager: Kim;
 let cfg: settings.Settings;
 let tray: Tray;
 let gone = false; // when true indicates app is shutting down
-let lastBuildDirectory = '';
 
 if (!Electron.app.requestSingleInstanceLock()) {
   gone = true;
@@ -118,23 +116,8 @@ Electron.app.whenReady().then(async() => {
     return;
   }
 
-  imageManager = new Kim();
-  interface KimImage {
-    imageName: string,
-    tag: string,
-    imageID: string,
-    size: string
-  }
-  imageManager.on('images-changed', (images: KimImage[]) => {
-    window.send('images-changed', images);
-  });
-  imageManager.on('readiness-changed', (state: boolean) => {
-    window.send('images-check-state', state);
-  });
-
   window.send('k8s-current-port', cfg.kubernetes.port);
   k8smanager.start(cfg.kubernetes).catch(handleFailure);
-  imageManager.start();
 
   // Set up protocol handler for app://
   // This is needed because in packaged builds we'll not be allowed to access
@@ -162,9 +145,7 @@ Electron.app.whenReady().then(async() => {
   });
   window.openPreferences();
 
-  imageManager.on('kim-process-output', (data: string, isStderr: boolean) => {
-    window.send('kim-process-output', data, isStderr);
-  });
+  setupKim();
 });
 
 Electron.app.on('second-instance', () => {
@@ -188,7 +169,6 @@ Electron.app.on('before-quit', async(event) => {
     handleFailure(ex);
   } finally {
     gone = true;
-    imageManager.stop();
     Electron.app.quit();
   }
 });
@@ -263,126 +243,6 @@ Electron.app.on('certificate-error', (event, webContents, url, error, certificat
 
   // eslint-disable-next-line node/no-callback-literal
   callback(false);
-});
-
-Electron.ipcMain.on('confirm-do-image-deletion', async(event, imageName, imageID) => {
-  const choice = Electron.dialog.showMessageBoxSync( {
-    message:   `Delete image ${ imageName }?`,
-    type:      'warning',
-    buttons:   ['Yes', 'No'],
-    defaultId: 1,
-    title:     `Delete image ${ imageName }`,
-    cancelId:  1
-  });
-
-  if (choice === 0) {
-    try {
-      const maxNumAttempts = 2;
-      // On macOS a second attempt is needed to actually delete the image.
-      // Probably due to a timing issue on the server part of kim, but not determined why.
-      // Leave this in for windows in case it can happen there too.
-      let i = 0;
-
-      for (i = 0; i < maxNumAttempts; i++) {
-        await imageManager.deleteImage(imageID);
-        await imageManager.refreshImages();
-        if (!imageManager.listImages().some(image => image.imageID === imageID)) {
-          break;
-        }
-        await util.promisify(setTimeout)(500);
-      }
-      if (i === maxNumAttempts) {
-        console.log(`Failed to delete ${ imageID } in ${ maxNumAttempts } tries`);
-      }
-      event.reply('kim-process-ended', 0);
-    } catch (err) {
-      Electron.dialog.showMessageBox({
-        message: `Error trying to delete image ${ imageName } (${ imageID }):\n\n ${ err.stderr } `,
-        type:    'error'
-      });
-    }
-  }
-});
-
-Electron.ipcMain.on('do-image-build', async(event, taggedImageName: string) => {
-  const options: any = {
-    title:      'Pick the build directory',
-    properties: ['openFile'],
-    message:    'Please select the Dockerfile to use (could have a different name)'
-  };
-
-  if (lastBuildDirectory) {
-    options.defaultPath = lastBuildDirectory;
-  }
-  const results = Electron.dialog.showOpenDialogSync(options);
-
-  if (results === undefined) {
-    return;
-  }
-  if (results.length !== 1) {
-    console.log(`Expecting exactly one result, got ${ results.join(', ') }`);
-
-    return;
-  }
-  const pathParts = path.parse(results[0]);
-  let code;
-
-  lastBuildDirectory = pathParts.dir;
-  try {
-    code = (await imageManager.buildImage(lastBuildDirectory, pathParts.base, taggedImageName)).code;
-    await imageManager.refreshImages();
-  } catch (err) {
-    code = err.code;
-    Electron.dialog.showMessageBox({
-      message: `Error trying to build ${ taggedImageName }:\n\n ${ err.stderr } `,
-      type:    'error'
-    });
-  }
-  event.reply('kim-process-ended', code);
-});
-
-Electron.ipcMain.on('do-image-pull', async(event, imageName) => {
-  let taggedImageName = imageName;
-  let code;
-
-  if (!imageName.includes(':')) {
-    taggedImageName += ':latest';
-  }
-  try {
-    code = (await imageManager.pullImage(taggedImageName)).code;
-    await imageManager.refreshImages();
-  } catch (err) {
-    code = err.code;
-    Electron.dialog.showMessageBox({
-      message: `Error trying to pull ${ taggedImageName }:\n\n ${ err.stderr } `,
-      type:    'error'
-    });
-  }
-  event.reply('kim-process-ended', code);
-});
-
-Electron.ipcMain.on('do-image-push', async(event, imageName, imageID, tag) => {
-  const taggedImageName = `${ imageName }:${ tag }`;
-  let code;
-
-  try {
-    code = (await imageManager.pushImage(taggedImageName)).code;
-  } catch (err) {
-    code = err.code;
-    Electron.dialog.showMessageBox({
-      message: `Error trying to push ${ taggedImageName }:\n\n ${ err.stderr } `,
-      type:    'error'
-    });
-  }
-  event.reply('kim-process-ended', code);
-});
-
-Electron.ipcMain.handle('images-fetch', (event) => {
-  return imageManager.listImages();
-});
-
-Electron.ipcMain.handle('images-check-state', () => {
-  return imageManager.isReady;
 });
 
 Electron.ipcMain.on('k8s-state', (event) => {
@@ -645,13 +505,9 @@ function newK8sManager() {
   const mgr = K8s.factory();
 
   mgr.on('state-changed', (state: K8s.State) => {
+    mainEvents.emit('k8s-check-state', mgr);
     tray.emit('k8s-check-state', state);
     window.send('k8s-check-state', state);
-    if (state === K8s.State.STARTED) {
-      imageManager.start();
-    } else {
-      imageManager.stop();
-    }
   });
 
   mgr.on('current-port-changed', (port: number) => {
